+++ packages/react/src/components/AnimatedHeader/components/AnimatedHeader/AnimatedHeader.tsx
@@ -7,7 +7,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 import PropTypes from 'prop-types';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { ReactNode, useEffect, useRef, useState } from 'react';
 import {
   Grid,
   Column,
@@ -43,12 +43,13 @@ export interface SelectedWorkspace {
 }
 
 export interface Tile {
-  href: string | null;
+  href?: string | null;
   id: string;
-  mainIcon: string | null;
+  mainIcon?: string | null;
   secondaryIcon?: string | null;
   subtitle?: string | null;
-  title: string | null;
+  title?: string | null;
+  customContent?: ReactNode | null;
 }
 
 export interface TileGroup {
@@ -61,8 +62,10 @@ export interface AnimatedHeaderProps {
   allTiles: TileGroup[];
   allWorkspaces?: SelectedWorkspace[];
   description?: string;
-  handleHeaderItemsToString: (item: TileGroup | null) => string;
-  handleWorkspaceItemsToString: (item: SelectedWorkspace | null) => string;
+  handleHeaderItemsToString?: (item: TileGroup | null) => string;
+  renderHeaderSelectedItem?: (item: TileGroup | null) => ReactNode;
+  handleWorkspaceItemsToString?: (item: SelectedWorkspace | null) => string;
+  renderWorkspaceSelectedItem?: (item: SelectedWorkspace | null) => ReactNode;
   headerAnimation?: object;
   headerStatic?: React.JSX.Element;
   productName?: string;
@@ -85,6 +88,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
   headerAnimation,
   headerStatic,
   productName = '[Product name]',
+  renderHeaderSelectedItem,
+  renderWorkspaceSelectedItem,
   selectedTileGroup,
   selectedWorkspace,
   setSelectedTileGroup,
@@ -180,8 +185,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
         <div className={`${blockClass}__lottie-animation--container`}>
           <div
             ref={animationContainer}
-            className={`${blockClass}__lottie-animation ${
-              !open && lottieCollapsed
+            className={`${blockClass}__lottie-animation${
+              !open ? ` ${lottieCollapsed}` : ''
             }`}></div>
         </div>
 
@@ -207,8 +212,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
             sm={4}
             md={8}
             lg={4}
-            className={`${blockClass}__left-area-container ${
-              !open && descriptionCollapsed
+            className={`${blockClass}__left-area-container${
+              !open ? ` ${descriptionCollapsed}` : ''
             }`}>
             {description && (
               <h2 className={`${blockClass}__description`}>{description}</h2>
@@ -238,8 +243,13 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                     hideLabel
                     type="inline"
                     items={allTiles}
-                    itemToString={(item) => handleHeaderItemsToString(item)}
                     onChange={(e) => setSelectedTileGroup(e)}
+                    {...(handleHeaderItemsToString
+                      ? { itemToString: handleHeaderItemsToString }
+                      : {})}
+                    {...(renderHeaderSelectedItem
+                      ? { renderSelectedItem: renderHeaderSelectedItem }
+                      : {})}
                   />
                 </div>
               )}
@@ -250,8 +260,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
           <Column sm={4} md={8} lg={12} className={`${blockClass}__content`}>
             {allWorkspaces && (
               <div
-                className={`${blockClass}__workspace--container ${
-                  !open && contentCollapsed
+                className={`${blockClass}__workspace--container${
+                  !open ? ` ${contentCollapsed}` : ''
                 }`}>
                 <Dropdown
                   id={`${blockClass}__workspace`}
@@ -262,8 +272,16 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                   hideLabel
                   type="inline"
                   items={allWorkspaces}
-                  itemToString={(item) => handleWorkspaceItemsToString(item)}
                   onChange={(e) => setSelectedWorkspace(e)}
+                  {...(handleWorkspaceItemsToString
+                    ? { itemToString: handleWorkspaceItemsToString }
+                    : {})}
+                  {...(renderWorkspaceSelectedItem
+                    ? { renderSelectedItem: renderWorkspaceSelectedItem }
+                    : {})}
+                  {...(selectedWorkspace
+                    ? { selectedItem: selectedWorkspace }
+                    : {})}
                 />
               </div>
             )}
@@ -279,7 +297,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                     secondaryIcon={tile.secondaryIcon}
                     title={tile.title}
                     subtitle={tile.subtitle}
-                    productName={productName}></BaseTile>
+                    productName={productName}
+                    customContent={tile.customContent}></BaseTile>
                 );
               })}
             </div>
@@ -356,6 +375,20 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
    * Provide current product name
    */
   productName: PropTypes.string,
+
+  /**
+   * Helper function passed to downshift that allows the library to render a
+   * selected item to as an arbitrary ReactNode. By default it uses standard Carbon renderer that renders only item.label text
+   * (Dropdown under description in header)
+   */
+  renderHeaderSelectedItem: PropTypes.func,
+
+  /**
+   * Helper function passed to downshift that allows the library to render a
+   * selected item to as an arbitrary ReactNode. By default it uses standard Carbon renderer that renders only item.label text
+   * (Dropdown related to workspace selection)
+   */
+  renderWorkspaceSelectedItem: PropTypes.func,
 
   /**
    * The tile group that is active in the header