+++ packages/web-components/src/components/empty-state/components/empty-state/src/empty-state.ts
@@ -8,7 +8,7 @@
  */
 
 import { LitElement } from 'lit';
-import { property } from 'lit/decorators.js';
+import { property, state } from 'lit/decorators.js';
 
 // @ts-ignore
 import styles from './empty-state.scss?inline';
@@ -20,16 +20,35 @@ export default class emptyState extends LitElement {
   /**
    * Empty state heading
    */
-  @property({ type: String })
+  @property({ type: String, reflect: true })
   title;
   /**
    * Empty state subtext
    */
-  @property({ type: String })
+  @property({ type: String, reflect: true })
   subtitle;
   /**
    * Empty state subtext
    */
-  @property()
+  @property({ type: String, reflect: true })
   size: 'sm' | 'lg' = 'lg';
+  /**
+   * Determines which predefined illustration will be displayed
+   */
+  @property()
+  kind?:
+    | 'error'
+    | 'noData'
+    | 'noTags'
+    | 'notFound'
+    | 'notifications'
+    | 'unauthorized';
+  /**
+   * Empty state illustration theme variations. To ensure you use the correct themed illustrations, you can conditionally specify light or dark based on your app's current theme value.
+   */
+  @property({ type: String, reflect: true })
+  illustrationTheme?: 'light' | 'dark';
+
+  @state()
+  hasIllustration?: boolean;
 }